Add operators for symbolic calculus
